better file reorg
